Load 3.3.582 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1237 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/lex.c
@@ -863,9 +863,9 @@ init_lexer(void)
     add_permanent_define("__INT_MAX__", -1, string_copy(mtext), MY_FALSE);
     sprintf(mtext, "(%ld)", PINT_MIN);
     add_permanent_define("__INT_MIN__", -1, string_copy(mtext), MY_FALSE);
-    sprintf(mtext, "(%g)", FLT_MAX);
+    sprintf(mtext, "(%g)", DBL_MAX);
     add_permanent_define("__FLOAT_MAX__", -1, string_copy(mtext), MY_FALSE);
-    sprintf(mtext, "(%g)", FLT_MIN);
+    sprintf(mtext, "(%g)", DBL_MIN);
     add_permanent_define("__FLOAT_MIN__", -1, string_copy(mtext), MY_FALSE);
 
     /* Add the permanent macro definitions given on the commandline */
